Add files via upload
First iteration of Enigma machine done in python. For the custom element we added a randomly generated key that would be XOR'd with the fully encrypted message to return the encrypted message in Hexadecimal. This makes the encoded message much harder to interpret and would make decrypting it much more difficult. We also made the plugboard to be randomized each time the code is run in order to simulate the ever changing plugboard code.